ML_EndToEnd_Projects/requirements.txt
# Example requirements.txt for all ML/GenAI projects
fastapi
uvicorn
pydantic
scikit-learn
xgboost
lightgbm
pandas
numpy
matplotlib
seaborn
plotly
joblib
mlflow
opencv-python
face_recognition
Pillow
nltk
spacy
transformers
torch
torchvision
torchaudio
torch-geometric
prophet
statsmodels
yfinance
ultralytics
networkx
rdkit
pytesseract
pdf2image
diffusers
